Added tenant identifiers to connection string vaults

+++ src/MJCZone.DapperMatic.WebApi/Vaults/ConnectionStringsDatabaseVault.cs
@@ -1,5 +1,4 @@
 using Dapper;
-
 using MJCZone.DapperMatic.Models;
 using MJCZone.DapperMatic.WebApi.Options;
 
@@ -15,6 +14,7 @@ public class ConnectionStringsDatabaseVault : ConnectionStringsVaultBase
     private readonly string _tableName = "web_connection_strings";
     private readonly string _nameColumn = "name";
     private readonly string _valueColumn = "value";
+    private readonly string _tenantIdentifierColumn = "tenant_identifier";
 
     /// <summary>
     /// Initializes a new instance of the <see cref="ConnectionStringsDatabaseVault"/> class.
@@ -79,6 +79,16 @@ public class ConnectionStringsDatabaseVault : ConnectionStringsVaultBase
             _valueColumn = valuecolumn.ToString()!;
         }
 
+        if (
+            vaultOptions
+                .Settings.ToDictionary(k => k.Key.ToLowerInvariant(), v => v.Value)
+                .TryGetValue("tenantidentifiercolumn", out var tenantidentifiercolumn)
+            && !string.IsNullOrWhiteSpace(tenantidentifiercolumn?.ToString())
+        )
+        {
+            _tenantIdentifierColumn = tenantidentifiercolumn.ToString()!;
+        }
+
         // initialize the database
         InitializeDatabase();
     }
@@ -87,11 +97,13 @@ public class ConnectionStringsDatabaseVault : ConnectionStringsVaultBase
     /// Gets the connection string for the specified name.
     /// </summary>
     /// <param name="connectionStringName">The name of the connection string to retrieve.</param>
+    /// <param name="tenantIdentifier">The tenant identifier.</param>
     /// <param name="cancellationToken">The cancellation token.</param>
     /// <returns>The encrypted connection string if found; otherwise, null.</returns>
     protected override async Task<string?> GetEncryptedConnectionStringAsync(
         string connectionStringName,
-        CancellationToken cancellationToken
+        string? tenantIdentifier = null,
+        CancellationToken cancellationToken = default
     )
     {
         using var connection = DatabaseConnectionFactory.GetDbConnection(
@@ -99,10 +111,21 @@ public class ConnectionStringsDatabaseVault : ConnectionStringsVaultBase
             _providerType
         );
 
+        string Query(string? tenantIdentifier) =>
+            string.IsNullOrWhiteSpace(tenantIdentifier)
+                ? "SELECT {_valueColumn} FROM {_tableName} WHERE {_nameColumn} = @name AND {_tenantIdentifierColumn} IS NULL"
+                : $"SELECT {_valueColumn} FROM {_tableName} WHERE {_nameColumn} = @name AND {_tenantIdentifierColumn} = @tenantIdentifier";
+
         return await connection
             .QueryFirstOrDefaultAsync<string?>(
-                $"SELECT {_valueColumn} FROM {_tableName} WHERE {_nameColumn} = @name",
-                new { name = connectionStringName }
+                Query(tenantIdentifier),
+                new
+                {
+                    name = connectionStringName,
+                    tenantIdentifier = string.IsNullOrWhiteSpace(tenantIdentifier)
+                        ? null
+                        : tenantIdentifier,
+                }
             )
             .ConfigureAwait(false);
     }
@@ -112,12 +135,14 @@ public class ConnectionStringsDatabaseVault : ConnectionStringsVaultBase
     /// </summary>
     /// <param name="connectionStringName">The name of the connection string to set.</param>
     /// <param name="encryptedConnectionString">The connection string value to set.</param>
+    /// <param name="tenantIdentifier">The tenant identifier.</param>
     /// <param name="cancellationToken">The cancellation token.</param>
     /// <returns><see cref="Task"/> representing the asynchronous operation.</returns>
     protected override async Task SetEncryptedConnectionStringAsync(
         string connectionStringName,
         string encryptedConnectionString,
-        CancellationToken cancellationToken
+        string? tenantIdentifier = null,
+        CancellationToken cancellationToken = default
     )
     {
         using var connection = DatabaseConnectionFactory.GetDbConnection(
@@ -125,19 +150,43 @@ public class ConnectionStringsDatabaseVault : ConnectionStringsVaultBase
             _providerType
         );
 
+        string UpdateStatement(string? tenantIdentifier) =>
+            string.IsNullOrWhiteSpace(tenantIdentifier)
+                ? $"UPDATE {_tableName} SET {_valueColumn} = @value WHERE {_nameColumn} = @name AND {_tenantIdentifierColumn} IS NULL"
+                : $"UPDATE {_tableName} SET {_valueColumn} = @value WHERE {_nameColumn} = @name AND {_tenantIdentifierColumn} = @tenantIdentifier";
+
         var updates = await connection
             .ExecuteAsync(
-                $"UPDATE {_tableName} SET {_valueColumn} = @value WHERE {_nameColumn} = @name",
-                new { name = connectionStringName, value = encryptedConnectionString }
+                UpdateStatement(tenantIdentifier),
+                new
+                {
+                    name = connectionStringName,
+                    value = encryptedConnectionString,
+                    tenantIdentifier = string.IsNullOrWhiteSpace(tenantIdentifier)
+                        ? null
+                        : tenantIdentifier,
+                }
             )
             .ConfigureAwait(false);
 
         if (updates == 0)
         {
+            string InsertStatement(string? tenantIdentifier) =>
+                string.IsNullOrWhiteSpace(tenantIdentifier)
+                    ? $"INSERT INTO {_tableName} ({_nameColumn}, {_valueColumn}) VALUES (@name, @value)"
+                    : $"INSERT INTO {_tableName} ({_nameColumn}, {_valueColumn}, {_tenantIdentifierColumn}) VALUES (@name, @value, @tenantIdentifier)";
+
             await connection
                 .ExecuteAsync(
-                    $"INSERT INTO {_tableName} ({_nameColumn}, {_valueColumn}) VALUES (@name, @value)",
-                    new { name = connectionStringName, value = encryptedConnectionString }
+                    InsertStatement(tenantIdentifier),
+                    new
+                    {
+                        name = connectionStringName,
+                        value = encryptedConnectionString,
+                        tenantIdentifier = string.IsNullOrWhiteSpace(tenantIdentifier)
+                            ? null
+                            : tenantIdentifier,
+                    }
                 )
                 .ConfigureAwait(false);
         }
@@ -147,11 +196,13 @@ public class ConnectionStringsDatabaseVault : ConnectionStringsVaultBase
     /// Deletes the connection string for the specified name.
     /// </summary>
     /// <param name="connectionStringName">The name of the connection string to delete.</param>
+    /// <param name="tenantIdentifier">The tenant identifier.</param>
     /// <param name="cancellationToken">The cancellation token.</param>
     /// <returns><see cref="Task"/> representing the asynchronous operation.</returns>
     protected override async Task DeleteEncryptedConnectionStringAsync(
         string connectionStringName,
-        CancellationToken cancellationToken
+        string? tenantIdentifier = null,
+        CancellationToken cancellationToken = default
     )
     {
         using var connection = DatabaseConnectionFactory.GetDbConnection(
@@ -159,10 +210,21 @@ public class ConnectionStringsDatabaseVault : ConnectionStringsVaultBase
             _providerType
         );
 
+        string DeleteStatement(string? tenantIdentifier) =>
+            string.IsNullOrWhiteSpace(tenantIdentifier)
+                ? $"DELETE FROM {_tableName} WHERE {_nameColumn} = @name AND {_tenantIdentifierColumn} IS NULL"
+                : $"DELETE FROM {_tableName} WHERE {_nameColumn} = @name AND {_tenantIdentifierColumn} = @tenantIdentifier";
+
         await connection
             .ExecuteAsync(
-                $"DELETE FROM {_tableName} WHERE {_nameColumn} = @name",
-                new { name = connectionStringName }
+                DeleteStatement(tenantIdentifier),
+                new
+                {
+                    name = connectionStringName,
+                    tenantIdentifier = string.IsNullOrWhiteSpace(tenantIdentifier)
+                        ? null
+                        : tenantIdentifier,
+                }
             )
             .ConfigureAwait(false);
     }
@@ -197,6 +259,14 @@ public class ConnectionStringsDatabaseVault : ConnectionStringsVaultBase
                             typeof(string),
                             length: 4096,
                             isNullable: false
+                        ),
+                        new DmColumn(
+                            null,
+                            _tableName,
+                            _tenantIdentifierColumn,
+                            typeof(string),
+                            length: 512,
+                            isNullable: true
                         ),
                     ]
                 )
